Adding support for ${param:FOO} syntax
