+++ src/extensions/paramHelper.ts
@@ -126,3 +126,14 @@ export async function promptForNewParams(
   }
   return currentParams;
 }
+
+export function readParamsFile(envFilePath: string): any {
+  try {
+    const buf = fs.readFileSync(path.resolve(envFilePath), "utf8");
+    return dotenv.parse(buf.toString().trim(), { debug: true });
+  } catch (err) {
+    throw new FirebaseError(`Error reading --test-params file: ${err.message}\n`, {
+      original: err,
+    });
+  }
+}
